Struct ccthw_ash_instance::PhysicalDevice

pub struct PhysicalDevice { /* private fields */ }

A Vulkan physical device along with its properties and requested features.

Physical devices are purely descriptive and can be cloned without concern for underlying GPU resources.

Implementations

impl PhysicalDevice

pub fn queue_family_properties(&self) -> &[QueueFamilyProperties]

Properties for all queue families supported by this device.

pub fn available_extensions(&self) -> &[ExtensionProperties]

The set of all extensions available on this device.

pub fn available_extension_names(&self) -> &[String]

The set of all extension names available on this device.

pub fn properties(&self) -> &PhysicalDeviceProperties

The properties for this physical device.

pub fn features(&self) -> &PhysicalDeviceFeatures

The features requested when picking this device.

pub fn name(&self) -> String

The physical device name from the device properties struct.

pub fn enumerate_supported_devices( instance: &VulkanInstance, required_features: &PhysicalDeviceFeatures ) -> InstanceResult<Vec<Self>>

Enumerate all physical devices which support the required featuers.

Params

instance - the Vulkan instance which will provide access to the physical devices.

features - the features the device must support.