Trait ccthw_ash_instance::VulkanHandle

pub trait VulkanHandle {
    type Handle;

    // Required method
    unsafe fn raw(&self) -> &Self::Handle;
}

Types which implement this trait can provide the raw Vulkan resource handle as provided by Ash.

Required Associated Types

type Handle

Required Methods

unsafe fn raw(&self) -> &Self::Handle

Get the raw resource handle.

Safety

Unsafe because ownership is not transferred when accessing the raw handle. The caller is responsible for ensuring no copies of the handle outlive the original instance.